// src/__generated__/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Athlete = {
  __typename?: 'Athlete';
  /** The list of attempts made by the athlete in various events. */
  attempts: Array<Attempt>;
  attemptsAggregate?: Maybe<AthleteAttemptAttemptsAggregationSelection>;
  attemptsConnection: AthleteAttemptsConnection;
  /** The birthday of the athlete. */
  birthday: Scalars['Date']['output'];
  /** The list of competitions in which the athlete participated. */
  competitions: Array<Competition>;
  competitionsAggregate?: Maybe<AthleteCompetitionCompetitionsAggregationSelection>;
  competitionsConnection: AthleteCompetitionsConnection;
  /** The timestamp when the athlete's profile was created. */
  createdAt: Scalars['DateTime']['output'];
  /** The list of data sources associated with the athlete. */
  dataSources: Array<DataSource>;
  dataSourcesAggregate?: Maybe<AthleteDataSourceDataSourcesAggregationSelection>;
  dataSourcesConnection: AthleteDataSourcesConnection;
  /** The list of events in which the athlete participated. */
  events: Array<Event>;
  eventsAggregate?: Maybe<AthleteEventEventsAggregationSelection>;
  eventsConnection: AthleteEventsConnection;
  /** The unique identifier of the athlete. */
  id: Scalars['ID']['output'];
  /** The full name of the athlete. Recorded as parsed. */
  name: Scalars['String']['output'];
  /** The list of nations represented by the athlete. */
  nations: Array<Nation>;
  nationsAggregate?: Maybe<AthleteNationNationsAggregationSelection>;
  nationsConnection: AthleteNationsConnection;
  /** The timestamp when the athlete's profile was last updated. */
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Indicates whether the athlete's profile is verified. */
  verified: Scalars['Boolean']['output'];
};
// ...
export type Attempt = {
  __typename?: 'Attempt';
  /** The athlete who made the attempt. */
  athlete: Athlete;
  athleteAggregate?: Maybe<AttemptAthleteAthleteAggregationSelection>;
  athleteConnection: AttemptAthleteConnection;
  /** The timestamp when the attempt was created. */
  createdAt: Scalars['DateTime']['output'];
  /** The date and time of the attempt. */
  dateTime?: Maybe<Scalars['DateTime']['output']>;
  /** The event in which the attempt was made. */
  event: Event;
  eventAggregate?: Maybe<AttemptEventEventAggregationSelection>;
  eventConnection: AttemptEventConnection;
  /** The unique identifier of the attempt. */
  id: Scalars['ID']['output'];
  /** The nation associated with the attempt. */
  nation: Nation;
  nationAggregate?: Maybe<AttemptNationNationAggregationSelection>;
  nationConnection: AttemptNationConnection;
  /** The number of the attempt by the athlete in the event. */
  number: Scalars['Int']['output'];
  /** Indicates whether the attempt was successful. */
  successful: Scalars['Boolean']['output'];
  /** The timestamp when the attempt was last updated. */
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Indicates whether the attempt is verified. */
  verified: Scalars['Boolean']['output'];
  /** The weight lifted or achieved in the attempt. */
  weight: Scalars['Float']['output'];
};
// ...
export type Competition = {
  __typename?: 'Competition';
  /** The list of athletes who participated in this competition. */
  athletes: Array<Athlete>;
  athletesAggregate?: Maybe<CompetitionAthleteAthletesAggregationSelection>;
  athletesConnection: CompetitionAthletesConnection;
  /** The timestamp when the competition was created. */
  createdAt: Scalars['DateTime']['output'];
  /** The list of data sources associated with this competition. */
  dataSources: Array<DataSource>;
  dataSourcesAggregate?: Maybe<CompetitionDataSourceDataSourcesAggregationSelection>;
  dataSourcesConnection: CompetitionDataSourcesConnection;
  /** The date of the competition. */
  date: Scalars['Date']['output'];
  /** The list of events in this competition. */
  events: Array<Event>;
  eventsAggregate?: Maybe<CompetitionEventEventsAggregationSelection>;
  eventsConnection: CompetitionEventsConnection;
  /** The unique identifier of the competition. */
  id: Scalars['ID']['output'];
  /** The name of the competition. */
  name: Scalars['String']['output'];
  /** The nation hosting the competition. */
  nation: Nation;
  nationAggregate?: Maybe<CompetitionNationNationAggregationSelection>;
  nationConnection: CompetitionNationConnection;
  /** Indicates whether data for this competition has been scraped. */
  scraped: Scalars['Boolean']['output'];
  /** The list of sports associated with this competition. */
  sports: Array<Sport>;
  sportsAggregate?: Maybe<CompetitionSportSportsAggregationSelection>;
  sportsConnection: CompetitionSportsConnection;
  /** The timestamp when the competition was last updated. */
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Indicates whether the competition is verified. */
  verified: Scalars['Boolean']['output'];
};
// ...
export type CompetitionDataSourcesConnection = {
  __typename?: 'CompetitionDataSourcesConnection';
  edges: Array<CompetitionDataSourcesRelationship>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type CompetitionDataSourcesRelationship = DataSourceEdge & {
  __typename?: 'CompetitionDataSourcesRelationship';
  /** The timestamp when this relationship was created. */
  createdAt: Scalars['DateTime']['output'];
  cursor: Scalars['String']['output'];
  /** The unique identifier of the entity associated with the data source. */
  entityId: Scalars['String']['output'];
  node: DataSource;
  /** The timestamp when this relationship was last updated. */
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
};
// ...
export type DataSource = {
  __typename?: 'DataSource';
  /** The list of athletes associated with this data source. */
  athletes: Array<Athlete>;
  athletesAggregate?: Maybe<DataSourceAthleteAthletesAggregationSelection>;
  athletesConnection: DataSourceAthletesConnection;
  /** The list of competitions associated with this data source. */
  competitions: Array<Competition>;
  competitionsAggregate?: Maybe<DataSourceCompetitionCompetitionsAggregationSelection>;
  competitionsConnection: DataSourceCompetitionsConnection;
  /** The timestamp when the data source's profile was created. */
  createdAt: Scalars['DateTime']['output'];
  /** The unique identifier of the data source. */
  id: Scalars['ID']['output'];
  /** The name of the data source. */
  name: Scalars['String']['output'];
  /** The timestamp when the data source's profile was last updated. */
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Indicates whether the data source's profile is verified. */
  verified: Scalars['Boolean']['output'];
};
// ...
export type Event = {
  __typename?: 'Event';
  /** The list of athletes who participated in this event. */
  athletes: Array<Athlete>;
  athletesAggregate?: Maybe<EventAthleteAthletesAggregationSelection>;
  athletesConnection: EventAthletesConnection;
  /** The list of attempts made in this event. */
  attempts: Array<Attempt>;
  attemptsAggregate?: Maybe<EventAttemptAttemptsAggregationSelection>;
  attemptsConnection: EventAttemptsConnection;
  /** The class of the event. */
  class?: Maybe<EventClass_Iwf>;
  /** The competition to which this event belongs. */
  competition: Competition;
  competitionAggregate?: Maybe<EventCompetitionCompetitionAggregationSelection>;
  competitionConnection: EventCompetitionConnection;
  /** The timestamp when the event was created. */
  createdAt: Scalars['DateTime']['output'];
  /** The date of the event. */
  date?: Maybe<Scalars['Date']['output']>;
  /** The gender class of the event */
  genderClass: Scalars['String']['output'];
  /** The unique identifier of the event. */
  id: Scalars['ID']['output'];
  /** The type of the event. */
  type: EventType_Iwf;
  /** The timestamp when the event was last updated. */
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Indicates whether the event is verified. */
  verified: Scalars['Boolean']['output'];
  /** The weight class of the event */
  weightClass: Scalars['String']['output'];
};
// ...
export type Nation = {
  __typename?: 'Nation';
  /** The list of athletes representing this nation. */
  athletes: Array<Athlete>;
  athletesAggregate?: Maybe<NationAthleteAthletesAggregationSelection>;
  athletesConnection: NationAthletesConnection;
  /** The code representing the nation (e.g., ISO code). */
  code: Scalars['String']['output'];
  /** The list of competitions associated with this nation. */
  competitions: Array<Competition>;
  competitionsAggregate?: Maybe<NationCompetitionCompetitionsAggregationSelection>;
  competitionsConnection: NationCompetitionsConnection;
  /** The timestamp when the nation's profile was created. */
  createdAt: Scalars['DateTime']['output'];
  /** The unique identifier of the nation. */
  id: Scalars['ID']['output'];
  /** The name of the nation. */
  name?: Maybe<Scalars['String']['output']>;
  /** The timestamp when the nation's profile was last updated. */
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Indicates whether the nation's profile is verified. */
  verified: Scalars['Boolean']['output'];
};
// ...
export type Query = {
  __typename?: 'Query';
  athletes: Array<Athlete>;
  athletesAggregate: AthleteAggregateSelection;
  athletesConnection: AthletesConnection;
  attempts: Array<Attempt>;
  attemptsAggregate: AttemptAggregateSelection;
  attemptsConnection: AttemptsConnection;
  competitions: Array<Competition>;
  competitionsAggregate: CompetitionAggregateSelection;
  competitionsConnection: CompetitionsConnection;
  dataSources: Array<DataSource>;
  dataSourcesAggregate: DataSourceAggregateSelection;
  dataSourcesConnection: DataSourcesConnection;
  events: Array<Event>;
  eventsAggregate: EventAggregateSelection;
  eventsConnection: EventsConnection;
  nations: Array<Nation>;
  nationsAggregate: NationAggregateSelection;
  nationsConnection: NationsConnection;
  sports: Array<Sport>;
  sportsAggregate: SportAggregateSelection;
  sportsConnection: SportsConnection;
};
// ...
export type Sport = {
  __typename?: 'Sport';
  /** The list of athletes who participated in this sport. */
  athletes: Array<Athlete>;
  athletesAggregate?: Maybe<SportAthleteAthletesAggregationSelection>;
  athletesConnection: SportAthletesConnection;
  /** The list of competitions associated with this sport. */
  competitions: Array<Competition>;
  competitionsAggregate?: Maybe<SportCompetitionCompetitionsAggregationSelection>;
  competitionsConnection: SportCompetitionsConnection;
  /** The timestamp when the sport's profile was created. */
  createdAt: Scalars['DateTime']['output'];
  /** The unique identifier of the sport. */
  id: Scalars['ID']['output'];
  /** The name of the sport. */
  name: Scalars['String']['output'];
  /** The timestamp when the sport's profile was last updated. */
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
  /** Indicates whether the sport's profile is verified. */
  verified: Scalars['Boolean']['output'];
};
// ...
export type DetailedCompetitionResultsQuery = { __typename?: 'Query', competitions: Array<{ __typename?: 'Competition', id: string, name: string, date: any, sports: Array<{ __typename?: 'Sport', name: string }>, dataSourcesConnection: { __typename?: 'CompetitionDataSourcesConnection', edges: Array<{ __typename?: 'CompetitionDataSourcesRelationship', entityId: string, node: { __typename?: 'DataSource', name: string } }> }, events: Array<{ __typename?: 'Event', id: string, genderClass: string, weightClass: string, attempts: Array<{ __typename?: 'Attempt', successful: boolean, weight: number, number: number, nation: { __typename?: 'Nation', code: string }, athlete: { __typename?: 'Athlete', birthday: any, name: string, verified: boolean, id: string, competitionsConnection: { __typename?: 'AthleteCompetitionsConnection', edges: Array<{ __typename?: 'AthleteCompetitionsRelationship', rank?: number | null, group: string, bodyweight: number }> } } }> }> }> };